Allow mysql root password to be supplied as an environment variable.
The envirnonment variable MYSQL_ROOT_PASSWORD can be set if required by
the mysql server. This is espacially usefull to run tests with a local server,
that requires a root password.

If no password is given, setup commands will run without a password.
So Travis CI should keep working as before.