Use CLIENT_DEPRECATE_EOF flag to make decision about EOF messages
Using server_version from handshake is incorrect to check that a server uses
EOF messages because all versions of MariaDB always have 5.5.5 as prefix
here even if they are marked EOF as deprecated.

This commit changes greeting behavior to use CLIENT_DEPRECATE_EOF flag
and now it should be possible to use MariaDB >= 10.2.